Integrate the libcmyth half of Ahmed Shihab's live tv patch from March 26.

+++ libs/libcmyth/timestamp.c
@@ -464,7 +464,7 @@ cmyth_datetime_to_string(char *str, cmyth_timestamp_t ts)
 			ts->timestamp_minute,
 			ts->timestamp_second);
 	cmyth_dbg(CMYTH_DBG_ERROR, "original timestamp string: %s \n",str);
-	sprintf(str,"%u",(unsigned int)t_datetime);
+	sprintf(str,"%lu",(unsigned long) t_datetime);
 	cmyth_dbg(CMYTH_DBG_ERROR, "time in seconds: %s \n",str);
 	
 	return 0;
